% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmd.R
\name{dmd_error}
\alias{dmd_error}
\title{DMD reconstruction error}
\usage{
dmd_error(object)
}
\arguments{
\item{object}{A dmd object.}
}
\value{
List with error metrics (rmse, mae, mape, relative_error).
}
\description{
DMD reconstruction error
}
\examples{
t <- seq(0, 10, length.out = 100)
X <- rbind(sin(t), cos(t))
d <- dmd(X, rank = 2, dt = t[2] - t[1])

# Reconstruction error metrics
dmd_error(d)
}
