% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmd.R
\name{dmd_reconstruct}
\alias{dmd_reconstruct}
\title{Reconstruct data from DMD modes}
\usage{
dmd_reconstruct(object, n_steps = NULL, modes = NULL)
}
\arguments{
\item{object}{A dmd object.}

\item{n_steps}{Number of time steps. Defaults to original length.}

\item{modes}{Integer vector of mode indices (1-based), or NULL for all.}
}
\value{
Numeric matrix.
}
\description{
Reconstruct data from DMD modes
}
\examples{
t <- seq(0, 10, length.out = 100)
X <- rbind(sin(t), cos(t))
d <- dmd(X, rank = 2, dt = t[2] - t[1])

recon <- dmd_reconstruct(d, n_steps = ncol(X))
dim(recon)
}
