% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmd.R
\name{dmd_residual}
\alias{dmd_residual}
\title{DMD residual analysis}
\usage{
dmd_residual(object)
}
\arguments{
\item{object}{A dmd object.}
}
\value{
List with residual_norm and residual_relative.
}
\description{
DMD residual analysis
}
\examples{
t <- seq(0, 10, length.out = 100)
X <- rbind(sin(t), cos(t))
d <- dmd(X, rank = 2, dt = t[2] - t[1])

dmd_residual(d)
}
