% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmd.R
\name{dmd_stability}
\alias{dmd_stability}
\title{DMD stability analysis}
\usage{
dmd_stability(object, tol = 1e-06)
}
\arguments{
\item{object}{A dmd object.}

\item{tol}{Tolerance for marginal classification.}
}
\value{
List with stability information.
}
\description{
DMD stability analysis
}
\examples{
t <- seq(0, 10, length.out = 100)
X <- rbind(sin(t), cos(t))
d <- dmd(X, rank = 2, dt = t[2] - t[1])

dmd_stability(d)
}
