% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{extended_standard_map}
\alias{extended_standard_map}
\title{Extended standard map (3D)}
\usage{
extended_standard_map(state, epsilon = 0.01, delta = 0.001)
}
\arguments{
\item{state}{Numeric vector of length 3.}

\item{epsilon}{Perturbation parameter.}

\item{delta}{Coupling parameter.}
}
\value{
Updated state vector.
}
\description{
Extended standard map (3D)
}
\examples{
extended_standard_map(c(0.1, 0.2, 0.3))
}
