% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gla.R
\name{gla_reconstruct}
\alias{gla_reconstruct}
\title{Reconstruct from GLA}
\usage{
gla_reconstruct(object, modes_to_use = NULL)
}
\arguments{
\item{object}{A gla object.}

\item{modes_to_use}{Integer vector of mode indices (1-based), or NULL.}
}
\value{
Numeric matrix.
}
\description{
Reconstruct from GLA
}
\examples{
t <- seq(0, 10, length.out = 200)
y <- rbind(sin(t), cos(t))
g <- gla(y, n_eigenvalues = 2)

recon <- gla_reconstruct(g)
dim(recon)
}
