% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankel.R
\name{hankel_reconstruct}
\alias{hankel_reconstruct}
\title{Reconstruct from Hankel-DMD}
\usage{
hankel_reconstruct(object, n_steps)
}
\arguments{
\item{object}{A hankel_dmd object.}

\item{n_steps}{Number of time steps.}
}
\value{
Numeric matrix.
}
\description{
Reconstruct from Hankel-DMD
}
\examples{
t <- seq(0, 4 * pi, length.out = 200)
y <- matrix(sin(t), nrow = 1)
h <- hankel_dmd(y, delays = 10)

recon <- hankel_reconstruct(h, 50)
}
