% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonic.R
\name{harmonic_time_average}
\alias{harmonic_time_average}
\title{Harmonic Time Average}
\usage{
harmonic_time_average(
  map_name,
  initial_condition,
  observable = "sin_pi",
  omega = 0.1,
  n_iter = 10000,
  ...
)
}
\arguments{
\item{map_name}{Character map name.}

\item{initial_condition}{Numeric vector.}

\item{observable}{Character observable name.}

\item{omega}{Numeric frequency.}

\item{n_iter}{Integer iterations.}

\item{...}{Map parameters.}
}
\value{
List with magnitude, phase, hta_re, hta_im.
}
\description{
Harmonic Time Average
}
\examples{
# Harmonic time average of an orbit of the Chirikov standard map
harmonic_time_average("standard", c(0.1, 0.2), "sin_pi", 0.1, 500)
}
