% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{henon_map}
\alias{henon_map}
\title{Henon map (2D dissipative)}
\usage{
henon_map(state, a = 1.4, b = 0.3)
}
\arguments{
\item{state}{Numeric vector c(x, y).}

\item{a}{Parameter a.}

\item{b}{Parameter b.}
}
\value{
Updated state vector.
}
\description{
Henon map (2D dissipative)
}
\examples{
henon_map(c(0, 0))

# Classic chaotic parameters
henon_map(c(0, 0), a = 1.4, b = 0.3)
}
