% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{logistic_map}
\alias{logistic_map}
\title{Logistic map (1D)}
\usage{
logistic_map(state, r = 3.9)
}
\arguments{
\item{state}{Numeric scalar.}

\item{r}{Growth rate parameter.}
}
\value{
Updated state value.
}
\description{
Logistic map (1D)
}
\examples{
logistic_map(0.5)

# In the chaotic regime
logistic_map(0.5, r = 3.9)
}
