% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmd.R
\name{predict.dmd}
\alias{predict.dmd}
\title{Predict from DMD model}
\usage{
\method{predict}{dmd}(object, n_ahead = 10, x0 = NULL, method = c("modes", "matrix"), ...)
}
\arguments{
\item{object}{A dmd object.}

\item{n_ahead}{Number of steps to predict.}

\item{x0}{Optional initial condition vector.}

\item{method}{"modes" (default) or "matrix".}

\item{...}{Additional arguments (ignored).}
}
\value{
Numeric matrix of predictions.
}
\description{
Predict from DMD model
}
\examples{
t <- seq(0, 10, length.out = 100)
X <- rbind(sin(t), cos(t))
d <- dmd(X, rank = 2, dt = t[2] - t[1])

# Forecast 10 steps beyond the input
pred <- predict(d, n_ahead = 10)
dim(pred)
}
