% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gla.R
\name{predict.gla}
\alias{predict.gla}
\title{Predict from GLA}
\usage{
\method{predict}{gla}(object, n_ahead = 10, ...)
}
\arguments{
\item{object}{A gla object.}

\item{n_ahead}{Number of steps to predict.}

\item{...}{Additional arguments (ignored).}
}
\value{
Numeric matrix.
}
\description{
Predict from GLA
}
\examples{
t <- seq(0, 10, length.out = 200)
y <- rbind(sin(t), cos(t))
g <- gla(y, n_eigenvalues = 2)

pred <- predict(g, n_ahead = 5)
}
