% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankel.R
\name{predict.hankel_dmd}
\alias{predict.hankel_dmd}
\title{Predict from Hankel-DMD}
\usage{
\method{predict}{hankel_dmd}(object, n_ahead = 10, ...)
}
\arguments{
\item{object}{A hankel_dmd object.}

\item{n_ahead}{Number of steps to predict.}

\item{...}{Additional arguments (ignored).}
}
\value{
Numeric matrix.
}
\description{
Predict from Hankel-DMD
}
\examples{
t <- seq(0, 4 * pi, length.out = 200)
y <- matrix(sin(t), nrow = 1)
h <- hankel_dmd(y, delays = 10)

pred <- predict(h, n_ahead = 10)
}
