% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{standard_map}
\alias{standard_map}
\title{Standard map (Chirikov)}
\usage{
standard_map(state, epsilon = 0.12)
}
\arguments{
\item{state}{Numeric vector c(x, y).}

\item{epsilon}{Perturbation parameter.}
}
\value{
Updated state vector.
}
\description{
Standard map (Chirikov)
}
\examples{
# One iteration from a given state
standard_map(c(0.1, 0.2))

# Stronger nonlinearity
standard_map(c(0.1, 0.2), epsilon = 0.5)
}
