% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{koopman.dmd-package}
\alias{koopman.dmd}
\alias{koopman.dmd-package}
\title{koopman.dmd: Koopman Operator and Dynamic Mode Decomposition for Dynamical Systems}
\description{
Dynamic Mode Decomposition (DMD) with Koopman operator theory extensions, powered by a Rust backend via 'extendr'. Provides standard DMD (Schmid, 2010, \doi{10.1017/S0022112010001217}), Extended DMD with lifting functions, Hankel-DMD via time-delay embedding, Generalized Laplace Analysis for direct eigenfunction computation, harmonic time averages, and mesochronic harmonic plots for phase space analysis (Mezic, 2020, \doi{10.48550/arXiv.2009.05883}). Includes built-in area-preserving and chaotic maps for experimentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimeharrisjr/rust-dmd}
  \item \url{https://jimeharrisjr.github.io/rust-dmd/}
  \item Report bugs at \url{https://github.com/jimeharrisjr/rust-dmd/issues}
}

}
\author{
\strong{Maintainer}: James Harris \email{jimeharrisjr@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item The authors of the dependency Rust crates (see inst/AUTHORS file for details) [contributor]
}

}
